requirements.txt
django==3.2.5
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-mptt==0.12.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
psycopg2-binary==2.9.1
